Support class-based components and callable signature detection (#56)
This PR represents a big set of changes to support class-based components, as described in #53. When merged, this resolves #53 .

The `README.md` is updated to reflect these changes, all tests are updated, and there are very many new tests.

For a basic description of these changes from a user's perspective, see [my comments to #53 here](https://github.com/t-strings/tdom/issues/53#issuecomment-3293837917). 

There is also a new `CallableInfo` facility: a frozen dataclass that holds on to _just_ the information we need when inspecting a callable (using `inspect.signature()` now) to understand how to invoke it. We do this (rather than holding on to `inspect.Signature` instances directly) for cachability; `get_callable_info(c: Callable)` sits behind an `@lru_cache()`.